Fixed link between input and titlepage
Now when we propmt some values to the input controller, it shows them immediately to the title page.
It has been done through a huge violation of the MVVM pattern, so it should be fixed in a next versions.
I should remove Singleton and replace it with some DataContext-related code.